requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.4.post0
    # via
    #   -r requirements.in
    #   discord.py
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via aiohttp
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via aiohttp
charset-normalizer==2.0.4
    # via requests
discord.py==1.7.3
    # via -r requirements.in
future==0.18.2
    # via python-twitter
greenlet==1.1.1
    # via sqlalchemy
idna==3.2
    # via
    #   requests
    #   yarl
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
oauthlib==3.1.1
    # via requests-oauthlib
pydantic==1.8.2
    # via sqlmodel
python-dotenv==0.19.0
    # via -r requirements.in
python-twitter==3.5
    # via -r requirements.in
requests==2.26.0
    # via
    #   -r requirements.in
    #   python-twitter
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via python-twitter
sqlalchemy==1.4.23
    # via sqlmodel
sqlalchemy2-stubs==0.0.2a14
    # via sqlmodel
sqlmodel==0.0.4
    # via -r requirements.in
typing-extensions==3.10.0.0
    # via
    #   aiohttp
    #   pydantic
    #   sqlalchemy2-stubs
urllib3==1.26.6
    # via requests
yarl==1.6.3
    # via aiohttp
